+++ packages/kit/src/exports/vite/preview/index.js
@@ -71,7 +71,19 @@ export async function preview(vite, vite_config, svelte_config) {
 
 		vite.middlewares.use((req, res, next) => {
 			const original_url = /** @type {string} */ (req.url);
-			const { pathname } = new URL(original_url, 'http://dummy');
+			const { pathname, search } = new URL(original_url, 'http://dummy');
+
+			// if `paths.base === '/a/b/c`, then the root route is `/a/b/c/`,
+			// regardless of the `trailingSlash` route option
+			if (base.length > 1 && pathname === base) {
+				let location = base + '/';
+				if (search) location += search;
+				res.writeHead(307, {
+					location
+				});
+				res.end();
+				return;
+			}
 
 			if (pathname.startsWith(base)) {
 				next();
@@ -102,7 +114,7 @@ export async function preview(vite, vite_config, svelte_config) {
 					return;
 				}
 
-				const { pathname } = new URL(/** @type {string} */ (req.url), 'http://dummy');
+				const { pathname, search } = new URL(/** @type {string} */ (req.url), 'http://dummy');
 
 				let filename = normalizePath(
 					join(svelte_config.kit.outDir, 'output/prerendered/pages' + pathname)
@@ -113,6 +125,7 @@ export async function preview(vite, vite_config, svelte_config) {
 					const has_trailing_slash = pathname.endsWith('/');
 					const html_filename = `${filename}${has_trailing_slash ? 'index.html' : '.html'}`;
 
+					/** @type {string | undefined} */
 					let redirect;
 
 					if (is_file(html_filename)) {
@@ -127,6 +140,7 @@ export async function preview(vite, vite_config, svelte_config) {
 					}
 
 					if (redirect) {
+						if (search) redirect += search;
 						res.writeHead(307, {
 							location: redirect
 						});