+++ packages/kit/src/runtime/client/constants.js
@@ -7,5 +7,6 @@ export const PRELOAD_PRIORITIES = /** @type {const} */ ({
 	hover: 2,
 	viewport: 3,
 	eager: 4,
-	off: -1
+	off: -1,
+	false: -1
 });