+++ packages/kit/test/apps/basics/src/routes/actions/form-errors/adjacent-error-boundary/+page.svelte
@@ -1,7 +1,7 @@
 <script>
-  import { enhance } from '$app/forms';
+	import { enhance } from '$app/forms';
 </script>
 
 <form method="POST" use:enhance>
-  <button type="submit">submit</button>
+	<button type="submit">submit</button>
 </form>